lots of updates to Dashboard.js and ReviewPage.js, removed a few lines from App.js
Changed ReviewPage.js from functional component to Class component. Even though functional component is newer way to write components, rest of this project uses class components, so to make sure everything works right (and to be consistent) made it a class component. This messed up some of the styling on ReviewPage but will fix it later. Removed route to /reviewpage in App.js. Now when clicking on a class in dashboard it goes straight to the reviewpage (mounts reviewpage component) without having to route to /reviewpage. Had to add a button on ReviewPage to go back to Dashboard. Dashboard mostly done (search bar left). Work still to be done to ReviewPage, but have successfully sent class that was clicked on as a prop to ReviewPage.